Optmizing server actions, update ui etc

+++ lib/invoice-actions.ts
@@ -4,6 +4,7 @@ import { revalidatePath } from "next/cache";
 import prisma from "@/lib/prisma";
 import { Prisma, Invoice, Order, PaymentStatus } from "@prisma/client";
 import { requireAuth } from "./auth-action";
+import { requireRateLimit } from "./ratelimit";
 
 export type InvoiceActionState = {
   success?: boolean;
@@ -15,7 +16,15 @@ export async function createInvoiceAction(
   prevState: InvoiceActionState,
   formData: FormData
 ): Promise<InvoiceActionState> {
-  await requireAuth();
+  const user = await requireAuth();
+  if (!user) {
+    throw new Error("Sign In");
+  }
+  await requireRateLimit({
+    windowSec: 60, // 1 minute window
+    max: 10, //10 uploads per minute
+    identifier: user.id,
+  });
   try {
     // Extract and parse data
     const invoiceNumber = formData.get("invoiceNumber") as string;
@@ -98,7 +107,15 @@ export async function updateInvoiceAction(
   prevState: InvoiceActionState,
   formData: FormData
 ): Promise<InvoiceActionState> {
-  await requireAuth();
+  const user = await requireAuth();
+  if (!user) {
+    throw new Error("Sign In");
+  }
+  await requireRateLimit({
+    windowSec: 60, // 1 minute window
+    max: 10, //10 uploads per minute
+    identifier: user.id,
+  });
   try {
     const invoiceNumber = formData.get("invoiceNumber") as string;
     const orderNumber = formData.get("orderNumber") as string;
@@ -180,7 +197,15 @@ export async function updateInvoiceAction(
 export async function deleteInvoiceAction(
   invoiceId: string
 ): Promise<InvoiceActionState> {
-  await requireAuth();
+  const user = await requireAuth();
+  if (!user) {
+    throw new Error("Sign In");
+  }
+  await requireRateLimit({
+    windowSec: 60, // 1 minute window
+    max: 10, //10 uploads per minute
+    identifier: user.id,
+  });
   try {
     await prisma.invoice.delete({ where: { id: invoiceId } });
 
@@ -200,7 +225,15 @@ export async function deleteInvoiceAction(
 }
 
 export async function createInvoiceFromOrder(order: Order): Promise<Invoice> {
-  await requireAuth();
+  const user = await requireAuth();
+  if (!user) {
+    throw new Error("Sign In");
+  }
+  await requireRateLimit({
+    windowSec: 60, // 1 minute window
+    max: 10, //10 uploads per minute
+    identifier: user.id,
+  });
   const invoiceDate = order.completedAt || order.createdAt;
   const dueDate = new Date(invoiceDate);
   dueDate.setDate(dueDate.getDate() + 30);
